refactor: StoreTest 패키지 이동 및 파일명 변경
